+++ JsonToMidiConverter/Models/Song/BrushStroke.cs
@@ -5,8 +5,8 @@ namespace JsonToMidiConverter.Models.Song;
 public class BrushStroke
 {
     public Direction Direction { get; set; }
-    public int Duration { get; set; }
-    public float Shift { get; set; }
+    public double Duration { get; set; }
+    public double Shift { get; set; }
 
     public BrushStroke Clone() => new()
     {
