+++ JsonToMidiConverter/Models/Song/Text.cs
@@ -8,8 +8,8 @@ public sealed class Text : ISerializable
     [JsonPropertyName("text")]
     public string Content { get; set; } = string.Empty;
 
-    [JsonConverter(typeof(NullToDefaultConverter<int>))]
-    public int Width { get; set; } = 0;
+    [JsonConverter(typeof(NullToDefaultConverter<double>))]
+    public double Width { get; set; } = 0;
 
     
     public Text Clone() => new()
